Renamed Object(s) as InteractiveElement(s), filled in basic InteractiveElement class
