+++ src/main/java/InteractiveElements/Minus.java
@@ -1,4 +1,4 @@
-package Objects;
+package InteractiveElements;
 
 public interface Minus {
     // deduct points if met by the Player
