+++ src/main/java/InteractiveElements/Movable.java
@@ -1,4 +1,4 @@
-package Objects;
+package InteractiveElements;
 
 public interface Movable {
 }
