--- src/main/java/Objects/Object.java
@@ -1,5 +1,0 @@
-package Objects;
-
-public class Object {
-    // Consider the construction of Player. This is the base class for all objects to be seeded on the board.
-}
