--- src/main/java/Objects/Plus.java
@@ -1,4 +1,0 @@
-package Objects;
-
-public interface Plus {
-}
